refactor: Seperated student and teacher Schedule
