+++ src/pages/s/[classID]/schedule.tsx
@@ -11,11 +11,7 @@ import { useEffect, useState } from "react";
 // SK Components
 import {
   Button,
-  Dialog,
-  DialogSection,
-  Dropdown,
   Header,
-  KeyboardInput,
   MaterialIcon,
   RegularLayout,
   Search,
@@ -26,188 +22,26 @@ import {
 
 // Components
 import Schedule from "@components/Schedule";
-import DiscardDraft from "@components/dialogs/DiscardDraft";
 import BrandIcon from "@components/icons/BrandIcon";
 
 // Types
-import { DialogProps } from "@utils/types/common";
 import { Role } from "@utils/types/person";
 import { Schedule as ScheduleType } from "@utils/types/schedule";
 
 // Backend
-import { addPeriodtoSchedule } from "@utils/backend/schedule";
 import { SubjectListItem } from "@utils/types/subject";
 import { nameJoiner } from "@utils/helpers/name";
 
-interface AddPeriodProps extends DialogProps {
-  onSubmit: (formData: FormData) => void;
-}
-
-const AddPeriod = ({
-  show,
-  onClose,
-  onSubmit,
-}: AddPeriodProps): JSX.Element => {
-  const { t } = useTranslation(["schedule", "common"]);
-  const locale = useRouter().locale == "en-US" ? "en-US" : "th";
-  const [showDiscard, setShowDiscard] = useState<boolean>(false);
-
-  // Form control
-  const [form, setForm] = useState({
-    subject: 1,
-    day: "1",
-    periodStart: "",
-    duration: "1",
-  });
-
-  function validateAndSend() {
-    // Pre-parse validation
-    if (!form.periodStart) return false;
-
-    const periodStart = parseInt(form.periodStart);
-    const duration = parseInt(form.duration);
-    let formData = new FormData();
-
-    // Validates
-    if (form.subject < 0) return false;
-    if (!form.day) return false;
-    if (periodStart < 0 || periodStart > 10) return false;
-    if (duration < 1 || duration > 10) return false;
-
-    // Appends to form data
-    formData.append("subject", form.subject.toString());
-    formData.append("day", form.day);
-    formData.append("period-start", form.periodStart);
-    formData.append("duration", form.duration);
-
-    // Send
-    onSubmit(formData);
-    addPeriodtoSchedule(formData);
-
-    return true;
-  }
-
-  return (
-    <>
-      <Dialog
-        type="large"
-        label="add-period"
-        title={t("dialog.add.title")}
-        actions={[
-          { name: t("dialog.add.action.cancel"), type: "close" },
-          { name: t("dialog.add.action.save"), type: "submit" },
-        ]}
-        show={show}
-        onClose={() => setShowDiscard(true)}
-        onSubmit={() => validateAndSend() && onClose()}
-      >
-        <DialogSection name={t("dialog.add.form.title")} isDoubleColumn>
-          <Dropdown
-            name="subject"
-            label={t("dialog.add.form.subject")}
-            options={[
-              {
-                value: 1,
-                label: {
-                  "en-US": "English 1",
-                  th: "ภาษาอังกฤษ 1",
-                }[locale],
-              },
-            ]}
-            onChange={(e: number) => setForm({ ...form, subject: e })}
-          />
-          <Dropdown
-            name="day"
-            label={t("dialog.add.form.day")}
-            options={[
-              {
-                value: "1",
-                label: t("datetime.day.1", { ns: "common" }),
-              },
-              {
-                value: "2",
-                label: t("datetime.day.2", { ns: "common" }),
-              },
-              {
-                value: "3",
-                label: t("datetime.day.3", { ns: "common" }),
-              },
-              {
-                value: "4",
-                label: t("datetime.day.4", { ns: "common" }),
-              },
-              {
-                value: "5",
-                label: t("datetime.day.5", { ns: "common" }),
-              },
-            ]}
-            defaultValue="1"
-            onChange={(e: string) => setForm({ ...form, day: e })}
-          />
-          <KeyboardInput
-            name="period-start"
-            type="number"
-            label={t("dialog.add.form.periodStart")}
-            onChange={(e: string) => setForm({ ...form, periodStart: e })}
-            attr={{
-              min: 1,
-              max: 10,
-            }}
-          />
-          <KeyboardInput
-            name="duration"
-            type="number"
-            label={t("dialog.add.form.duration")}
-            defaultValue="1"
-            onChange={(e: string) => setForm({ ...form, duration: e })}
-            attr={{
-              min: 1,
-              max: 10,
-            }}
-          />
-        </DialogSection>
-      </Dialog>
-      <DiscardDraft
-        show={showDiscard}
-        onClose={() => setShowDiscard(false)}
-        onSubmit={() => {
-          setShowDiscard(false);
-          onClose();
-        }}
-      />
-    </>
-  );
-};
-
 const ScheduleSection = ({
-  role,
   schedule,
-  setShowAddPeriod,
 }: {
-  role: Role;
   schedule: ScheduleType;
-  setShowAddPeriod: Function;
 }): JSX.Element => {
   const { t } = useTranslation("schedule");
 
   return (
     <Section>
       <Schedule schedule={schedule} role="teacher" />
-      {role == "teacher" && (
-        <div className="flex flex-row items-center justify-end gap-2">
-          <Button
-            label={t("schedule.action.edit")}
-            type="outlined"
-            onClick={() => setShowAddPeriod(true)}
-          />
-          <Button
-            label={t("schedule.action.add")}
-            type="filled"
-            icon={<MaterialIcon icon="add" />}
-            onClick={() => setShowAddPeriod(true)}
-          />
-        </div>
-      )}
     </Section>
   );
 };
@@ -319,60 +153,12 @@ const SubjectListSection = ({
   );
 };
 
-const Subjects: NextPage<{
+const StudentSchedule: NextPage<{
   role: Role;
   schedule: ScheduleType;
   subjectList: Array<SubjectListItem>;
-}> = ({ role, schedule: fetchedSchedule, subjectList }) => {
+}> = ({ schedule, subjectList }) => {
   const { t } = useTranslation("schedule");
-  const [schedule, setSchedule] = useState<ScheduleType>(fetchedSchedule);
-  const [showAddPeriod, setShowAddPeriod] = useState<boolean>(false);
-
-  function addSchedulePeriod(formData: FormData) {
-    const day = parseInt(formData.get("day")?.toString() || "-1");
-    const periodStart = parseInt(
-      formData.get("period-start")?.toString() || "-1"
-    );
-    const duration = parseInt(formData.get("duration")?.toString() || "-1");
-
-    setSchedule({
-      content: schedule.content.map((scheduleRow) => {
-        if (scheduleRow.day == day) {
-          // Replace the Period with the `periodStart` in question
-          return {
-            // Keep Day the same
-            ...scheduleRow,
-
-            content: scheduleRow.content
-              // Remove the old Periods that overlap this new Period
-              .filter(
-                (schedulePeriod) =>
-                  schedulePeriod.periodStart + schedulePeriod.duration - 1 <
-                    periodStart ||
-                  schedulePeriod.periodStart >= periodStart + duration
-              )
-              // Append the new Period
-              .concat([
-                {
-                  periodStart,
-                  duration,
-                  // TODO: Fetch this
-                  subject: {
-                    name: {
-                      "en-US": { name: "New Period" },
-                      th: { name: "คาบสอนใหม่" },
-                    },
-                    teachers: [],
-                  },
-                },
-              ]),
-          };
-        } else {
-          return scheduleRow;
-        }
-      }),
-    });
-  }
 
   return (
     <>
@@ -386,18 +172,9 @@ const Subjects: NextPage<{
           />
         }
       >
-        <ScheduleSection
-          role={role}
-          schedule={schedule}
-          setShowAddPeriod={setShowAddPeriod}
-        />
+        <ScheduleSection schedule={schedule} />
         <SubjectListSection subjectList={subjectList} />
       </RegularLayout>
-      <AddPeriod
-        show={showAddPeriod}
-        onClose={() => setShowAddPeriod(false)}
-        onSubmit={(formData: FormData) => addSchedulePeriod(formData)}
-      />
     </>
   );
 };
@@ -406,7 +183,6 @@ export const getServerSideProps: GetServerSideProps = async ({
   locale,
   params,
 }) => {
-  const role: Role = "teacher";
   const schedule: ScheduleType = {
     content: [
       {
@@ -562,11 +338,10 @@ export const getServerSideProps: GetServerSideProps = async ({
         "common",
         "schedule",
       ])),
-      role,
       schedule,
       subjectList,
     },
   };
 };
 
-export default Subjects;
+export default StudentSchedule;
